refactor deployment scripts, move scripts

+++ scripts/edge_setup/deploy/add-module-to-deployment.sh
@@ -1,5 +1,6 @@
 # Copyright (c) Microsoft. All rights reserved.
 # Licensed under the MIT license. See LICENSE file in the project root for full license information.
+set -o pipefail
 
 script_dir=$(cd "$(dirname "$0")" && pwd)
 source ${script_dir}/_deployment-helpers.sh
@@ -11,7 +12,7 @@ MODULE_IMAGE_NAME=$4
 CREATE_OPTIONS=$5
 
 if [ "${CREATE_OPTIONS}" == "" ]; then
-    CREATE_OPTIONS="{}"
+    CREATE_OPTIONS="\"{}\""
 fi
 
 
@@ -20,6 +21,9 @@ if [ "${HUB_NAME}" == "" ] || [ "${DEVICE_ID}" == "" ] || [ "${MODULE_NAME}" == 
     echo    hubName is without '.azure-devices.net' suffix
     exit 1
 fi
+
+echo Creating manifest json with module ${MODULE_NAME}
+TEMPFILE=$(mktemp)
 
 #
 # JSON with required module JSON
@@ -37,19 +41,18 @@ read -d '' EMPTY_MODULE_JSON << EOF
 }
 EOF
 
-TEMPFILE=$(mktemp)
-
 BASE=$(az iot edge export-modules --device-id ${DEVICE_ID} --hub-name ${HUB_NAME} --query content)
 [ $? -eq 0 ] || { echo "az iot edge export-modules failed"; exit 1; }
 
-echo ${BASE} | jq . - \
+echo ${BASE} | jq . \
     | jq "${PATH_MODULES}.${MODULE_NAME} = ${EMPTY_MODULE_JSON}" \
     | jq "${PATH_MODULES}.${MODULE_NAME}.settings.image = \"${MODULE_IMAGE_NAME}\"" \
-    | jq "${PATH_MODULES}.${MODULE_NAME}.settings.createOptions = \"${CREATE_OPTIONS}\"" \
+    | jq "${PATH_MODULES}.${MODULE_NAME}.settings.createOptions = ${CREATE_OPTIONS}" \
     > ${TEMPFILE}
+[ $? -eq 0 ] || { "jq failed"; exit 1; }  
 
-echo Adding private registry to deployment manifest
-az iot edge set-modules --device-id ${DEVICE_ID} --hub-name ${HUB_NAME} --content ${TEMPFILE}
+echo Applying manifest json with module ${MODULE_NAME}
+az iot edge set-modules --device-id ${DEVICE_ID} --hub-name ${HUB_NAME} --content ${TEMPFILE} > /dev/null
 [ $? -eq 0 ] || { echo "az iot edge set-modules failed"; exit 1; }
 
 rm ${TEMPFILE} || true
